iot_integrations/requirements.txt
Django
djangorestframework
django-mqtt
